Fix what the first compile found, and make the next failure readable
Four things the headers settle and I had guessed: API_Guid.hpp carries
APIGuidToString and ACAPinc.h does not reach it, RSGetIndString is in the
RS module rather than in the kit's Inc, BMAllocateHandle needs BM.hpp,
and API_Attr_Head::flags is a short so a bitwise result has to be
narrowed back explicitly under /W3 /WX.

The build log needs admin rights on the repository to read, which nobody
fixing a compile error necessarily has. So the errors now become
annotations and a job summary, and the log is kept as an artifact.

+++ archicad-addon/Src/ProjectCommands.cpp
@@ -421,16 +421,16 @@ GS::ObjectState ModifyLayersCommand::Execute (const GS::ObjectState& parameters,
 			// rewrote all of it every time.
 			bool flag = false;
 			if (one.Get ("isHidden", flag)) {
-				layer.header.flags = flag ? (layer.header.flags | APILay_Hidden)
-										  : (layer.header.flags & ~APILay_Hidden);
+				layer.header.flags = static_cast<short> (flag ? (layer.header.flags | APILay_Hidden)
+														   : (layer.header.flags & ~APILay_Hidden));
 			}
 			if (one.Get ("isLocked", flag)) {
-				layer.header.flags = flag ? (layer.header.flags | APILay_Locked)
-										  : (layer.header.flags & ~APILay_Locked);
+				layer.header.flags = static_cast<short> (flag ? (layer.header.flags | APILay_Locked)
+														   : (layer.header.flags & ~APILay_Locked));
 			}
 			if (one.Get ("isWireframe", flag)) {
-				layer.header.flags = flag ? (layer.header.flags | APILay_ForceToWire)
-										  : (layer.header.flags & ~APILay_ForceToWire);
+				layer.header.flags = static_cast<short> (flag ? (layer.header.flags | APILay_ForceToWire)
+														   : (layer.header.flags & ~APILay_ForceToWire));
 			}
 
 			step = ACAPI_Attribute_Modify (&layer, nullptr);
